Implementation of core point determination

+++ src/algorithmMain.jl
@@ -162,8 +162,8 @@ function solve(
             end
 
             # Set-up counter for Benders cuts (Chen & Luedtke approach)
-            node.ext[:Benders_cuts_original] = Tuple{Int16, Symbol}[]
-            node.ext[:Benders_cuts_binary] = Tuple{Int16, Symbol}[]
+            node.ext[:Benders_cuts_original] = Tuple{Int64, Symbol}[]
+            node.ext[:Benders_cuts_binary] = Tuple{Int64, Symbol}[]
 
         end
     end
